refactor(dataset): own frontend dataset domain contracts
