--- yak-ops-ui/src/components/analysis/dataset-service.ts
@@ -1,5 +1,0 @@
-/** @deprecated Import Dataset HTTP operations from `@/services/dataset` instead. */
-export {
-  fetchAnalysisDatasets,
-  queryAnalysisDataset,
-} from '@/services/dataset';
