+++ yak-ops-ui/src/services/dataset/index.ts
@@ -1,3 +1,18 @@
 export * from './api';
 export * from './constants';
+export type {
+  Aggregation,
+  DatasetField,
+  DatasetFieldRole as DatasetFieldSemanticRole,
+  DatasetFieldType,
+  DatasetQueryColumn,
+  DatasetQueryColumnBinding,
+  DatasetQueryFilter,
+  DatasetQueryMetric,
+  DatasetQueryPayload,
+  DatasetQueryResult,
+  DatasetQuerySort,
+  PublishedDataset,
+  Scalar,
+} from './model';
 export type * from './types';
