+++ yak-ops-ui/src/services/dataset/model.ts
@@ -1,0 +1,82 @@
+export type DatasetFieldRole = 'dimension' | 'metric';
+export type DatasetFieldType = 'string' | 'number' | 'date' | 'datetime' | 'boolean' | 'unknown';
+export type Scalar = string | number | boolean | null;
+
+export interface DatasetField {
+  key: string;
+  label: string;
+  physicalName: string;
+  dataType: DatasetFieldType;
+  role: DatasetFieldRole;
+  nullable: boolean;
+  description?: string;
+}
+
+export interface PublishedDataset {
+  id: string;
+  name: string;
+  description: string;
+  status: 'ONLINE' | 'OFFLINE';
+  sourceTaskId: string;
+  sourceTaskName: string;
+  currentVersionNo?: number;
+  updatedAt: string;
+  fields: DatasetField[];
+}
+
+export type Aggregation = 'SUM' | 'AVG' | 'COUNT' | 'COUNT_DISTINCT' | 'MAX' | 'MIN';
+
+export interface DatasetQueryMetric {
+  fieldId: string;
+  aggregation: Aggregation;
+}
+
+export interface DatasetQueryFilter {
+  fieldId: string;
+  operator: 'EQ' | 'NE' | 'GT' | 'GTE' | 'LT' | 'LTE' | 'LIKE';
+  value: Scalar;
+}
+
+export interface DatasetQuerySort {
+  fieldId: string;
+  aggregation?: Aggregation;
+  direction: 'ASC' | 'DESC';
+}
+
+export interface DatasetQueryPayload {
+  dimensions: string[];
+  metrics: DatasetQueryMetric[];
+  filters: DatasetQueryFilter[];
+  sorts: DatasetQuerySort[];
+  limit: number;
+  timeoutSeconds: number;
+}
+
+export interface DatasetQueryColumnBinding {
+  key: string;
+  fieldId: string;
+  displayName: string;
+  dataType: string;
+  aggregation?: Aggregation | null;
+}
+
+export interface DatasetQueryColumn {
+  name: string;
+  label: string;
+  typeName: string;
+  jdbcType: number;
+  nullable: boolean;
+}
+
+export interface DatasetQueryResult {
+  queryId?: string;
+  datasetId: string;
+  datasetVersionId: string;
+  datasetVersionNo: number;
+  bindings: DatasetQueryColumnBinding[];
+  columns: DatasetQueryColumn[];
+  rows: Scalar[][];
+  returnedRows: number;
+  truncated: boolean;
+  elapsedMillis: number;
+}
