+++ yak-ops-ui/src/services/dataset/types.ts
@@ -72,13 +72,3 @@ export interface DatasetDetailWire {
   versions: DatasetVersionWire[];
   fields: DatasetFieldWire[];
 }
-
-export type {
-  Aggregation,
-  DatasetField,
-  DatasetFieldType,
-  DatasetQueryPayload,
-  DatasetQueryResult,
-  PublishedDataset,
-  Scalar,
-} from '@/components/analysis/model';
